# Conflicts:
#	app/DoctrineMigrations/Version144.php

+++ src/AppBundle/Controller/Report/DocumentController.php
@@ -41,7 +41,7 @@ class DocumentController extends RestController
     /**
      * GET document by id
      *
-     * @Route("/document/{id}")
+     * @Route("/document/{id}", requirements={"id":"\d+"})
      * @Method({"GET"})
      */
     public function getOneById(Request $request, $id)
@@ -61,7 +61,8 @@ class DocumentController extends RestController
 
 
     /**
-     * Delete document.
+     * Soft Delete document.
+     * Accessible only from deputy area
      *
      * @Method({"DELETE"})
      *
@@ -71,25 +72,70 @@ class DocumentController extends RestController
      *
      * @return array
      */
-    public function delete($id)
+    public function softDelete($id)
     {
-        $this->get('logger')->debug('Deleting document ' . $id);
+        /** @var $document EntityDir\Report\Document */
+        $document = $this->findEntityBy(EntityDir\Report\Document::class, $id);
 
-        try {
+        // enable if the check above is removed and the note is available for editing for the whole team
+        $this->denyAccessIfClientDoesNotBelongToUser($document->getReport()->getClient());
 
-            /** @var $document EntityDir\Report\Document */
-            $document = $this->findEntityBy(EntityDir\Report\Document::class, $id);
+        $this->getEntityManager()->remove($document);
 
-            // enable if the check above is removed and the note is available for editing for the whole team
-            $this->denyAccessIfClientDoesNotBelongToUser($document->getReport()->getClient());
-
-            $this->getEntityManager()->remove($document);
-
-            $this->getEntityManager()->flush();
-        } catch (\Exception $e) {
-            $this->get('logger')->error('Failed to delete document ID: ' . $id . ' - ' . $e->getMessage());
-        }
+        $this->getEntityManager()->flush();
 
         return [];
     }
+
+    /**
+     * Hard Delete
+     * Currently only accessed by admin area cron (no user login needed)
+     * Throw exception if a non-soft deleted document is asked for deletiong
+     *
+     * @Method({"DELETE"})
+     * @Route("/document/hard-delete/{id}")
+     *
+     * @param int $id
+     */
+    public function hardDelete(Request $request, $id)
+    {
+        if (!$this->getAuthService()->isSecretValidForRole(EntityDir\User::ROLE_ADMIN, $request)) {
+            throw new \RuntimeException('Endpoint only accessible from ADMIN container.', 403);
+        }
+
+        /* @var $repo EntityDir\Repository\DocumentRepository */
+        $repo = $this->getRepository(EntityDir\Report\Document::class);
+        /* @var $document EntityDir\Report\Document */
+        $document = $repo->findUnfilteredOneBy(['id'=>$id]);
+        if (!$document->getDeletedAt()) {
+            throw new \RuntimeException("Can't hard delete document $id, as it's not soft-deleted");
+        }
+
+        $this->getEntityManager()->remove($document);
+        $this->getEntityManager()->flush($document);
+
+        return $id;
+    }
+
+    /**
+     * GET soft-documents
+     * Currently only accessed by admin area cron (no user login needed)
+     *
+     * @Route("/document/soft-deleted")
+     * @Method({"GET"})
+     */
+    public function getSoftDeletedDocuments(Request $request)
+    {
+        if (!$this->getAuthService()->isSecretValidForRole(EntityDir\User::ROLE_ADMIN, $request)) {
+            throw new \RuntimeException('Endpoint only accessible from ADMIN container.', 403);
+        }
+
+        $this->setJmsSerialiserGroups(['document-id', 'document-storage-reference']);
+
+        /* @var $repo EntityDir\Repository\DocumentRepository */
+        $repo = $this->getRepository(EntityDir\Report\Document::class);
+
+        return $repo->retrieveSoftDeleted();
+    }
+
 }
